requirements.txt
yt-dlp>=2024.1.0
deep-translator>=1.11.0
bilibili-api-python>=16.0.0
